+++ RecipeManager/RecipeManager.cs
@@ -12,7 +12,7 @@ namespace RecipeManager
     {
         public const string PluginGUID = "MidnightsFX.RecipeManager";
         public const string PluginName = "RecipeManager";
-        public const string PluginVersion = "0.1.3";
+        public const string PluginVersion = "0.3.1";
 
         public Common.Config cfg;